Add ASGI event handlers (#30)
